Desafio incorporando librerias

+++ root/main.js
@@ -15,24 +15,10 @@ class vinilos {
 let discos = [];
 
 discos.push(
-  new vinilos(
-    "img/tapas/d12.jpg",
-    "George Harrison",
-    "living in the material world",
-    1973,
-    6000,
-    1
-  )
+  new vinilos("img/tapas/d12.jpg", "George Harrison", "living in the material world", 1973, 6000, 1)
 );
 discos.push(
-  new vinilos(
-    "img/tapas/d3.jpg",
-    "Pescado rabioso",
-    "pescado rabioso 2",
-    1973,
-    7000,
-    2
-  )
+  new vinilos("img/tapas/d3.jpg", "Pescado rabioso", "pescado rabioso 2", 1973, 7000, 2)
 );
 discos.push(
   new vinilos("img/tapas/d1.jpg", "David Bowie", "the next day", 2013, 5000, 3)
@@ -47,34 +33,13 @@ discos.push(
   new vinilos("img/tapas/d6.jpg", "Radiohead", "ok computer", 1997, 6500, 6)
 );
 discos.push(
-  new vinilos(
-    "img/tapas/d5.jpg",
-    "Pink floyd",
-    "the dark side of the moon",
-    1973,
-    6900,
-    7
-  )
+  new vinilos("img/tapas/d5.jpg", "Pink floyd", "the dark side of the moon", 1973, 6900, 7)
 );
 discos.push(
-  new vinilos(
-    "img/tapas/d8.jpg",
-    "Led zeppelin",
-    "houses of the holy",
-    1973,
-    6800,
-    8
-  )
+  new vinilos("img/tapas/d8.jpg", "Led zeppelin", "houses of the holy", 1973, 6800, 8)
 );
 discos.push(
-  new vinilos(
-    "img/tapas/d10.jpg",
-    "Luis Alberto Spinetta",
-    "para los arboles",
-    2003,
-    7100,
-    9
-  )
+  new vinilos("img/tapas/d10.jpg", "Luis Alberto Spinetta", "para los arboles", 2003, 7100, 9)
 );
 discos.push(
   new vinilos("img/tapas/d2.jpg", "Anderson.Paak", "malibu", 2016, 5500, 10)
@@ -87,40 +52,52 @@ discos.push(
 );
 console.log(discos);
 
-//-------------------------------------------------------------
+//--------------------------------------------------------------------------
 //Registro de usuario para encargar discos
 const registroUsuarios = [];
 
 const usuario = document.getElementById("usuario");
 const email = document.getElementById("email");
-
 const datos = document.getElementById("form1");
 
 datos.addEventListener("submit", (evento) => {
   evento.preventDefault();
 
-  const valorUsuario = usuario.value;
-  const valorEmail = email.value;
+  if (usuario.value === "" || email.value === "") {
+    Swal.fire({
+      icon: "warning",
+      title: "Falto completar algun campo",
+      text: "Ingrese todos sus datos personales",
+    });
+  } else {
+    const valorUsuario = usuario.value;
+    const valorEmail = email.value;
 
-  const usuarios = {
-    Nombre: valorUsuario,
-    Correo: valorEmail,
-  };
+    const usuarios = {
+      Nombre: valorUsuario,
+      Correo: valorEmail,
+    };
+    registroUsuarios.push(usuarios);
+    localStorage.setItem("usuarios", JSON.stringify(usuarios));
 
-  registroUsuarios.push(usuarios);
-
-  localStorage.setItem("usuarios", JSON.stringify(usuarios));
-
-  console.log(registroUsuarios);
+    Swal.fire({
+      position: "top-end",
+      icon: "success",
+      title: "Usuario registrado",
+      showConfirmButton: false,
+      timer: 1100,
+    });
+    console.log(registroUsuarios);
+  }
 });
 
-//discos a encargar
+//--------------------------------------------------------------------------
+//para encargar discos
 const encargues = [];
 
 const inputArtista = document.getElementById("input-artista");
 const inputAlbum = document.getElementById("input-album");
-const inputAnio = document.getElementById("input-anio");
-
+const inputAnio = Number(document.getElementById("input-anio"));
 const formulario = document.getElementById("form2");
 
 formulario.addEventListener("submit", (e) => {
@@ -131,9 +108,11 @@ formulario.addEventListener("submit", (e) => {
   const anio = inputAnio.value;
 
   encargues.push(new vinilos("img/disc1.png", art, album, anio, "a definir"));
+  console.log(encargues);
   formulario.reset();
 });
 
+//--------------------------------------------------------------------------
 //boton finalizar
 const pedidoUsuario = [];
 
@@ -142,24 +121,43 @@ const finalizar = document.getElementById("finalizar");
 finalizar.addEventListener("click", (ev) => {
   ev.preventDefault();
 
-  registroUsuarios.push(encargues);
-  pedidoUsuario.push(registroUsuarios);
+  if (registroUsuarios.length > 0 && encargues.length > 0) {
+    registroUsuarios.push(encargues);
+    pedidoUsuario.push(registroUsuarios);
 
-  console.log(pedidoUsuario);
+    console.log(pedidoUsuario);
+    console.log(registroUsuarios);
 
-  mensajeCargado(registroUsuarios);
+    mensajeCargado(registroUsuarios);
+  } else {
+    Swal.fire({
+      icon: "warning",
+      title: "No ha completado su pedido!",
+      text: "Registre sus datos y encargue al menos un disco",
+    });
+  }
 });
 
-//Mensaje de finalizado
+//Funcion para mensaje de finalizado
 const mensajeCargado = () => {
-  const mensajeFinal = document.getElementById("mensajeModal");
-  mensajeFinal.className = "mensaje-pedido";
-  mensajeFinal.innerHTML = `<b>${registroUsuarios[0].Nombre}</b>,
-  usted ha encargado ${encargues.length} discos.<br>
-  Pronto le enviaremos un email.`;
+  Swal.fire({
+    title: "Muchas gracias!",
+    imageUrl: "img/disc-foto.webp",
+    imageWidth: 400,
+    imageHeight: 200,
+    imageAlt: "Custom image",
+    html: `<b>${registroUsuarios[0].Nombre}</b>,
+    usted ha encargado ${encargues.length} disco/s.<br>
+    Pronto le enviaremos un email.`,
+    showCloseButton: true,
+    showCancelButton: false,
+    focusConfirm: true,
+    confirmButtonText: '<i class="fa fa-thumbs-up"></i> Listo',
+    confirmButtonAriaLabel: "Thumbs up, great!",
+  });
 };
 
-//------------------------------------------------------------------
+//--------------------------------------------------------------------------
 //armado de grilla con objetos del array
 
 const carrito = [];
@@ -167,6 +165,7 @@ const carrito = [];
 const productos = document.getElementById("galProd");
 const vaciar = document.getElementById("vaciar");
 const addCarrito = document.getElementById("add");
+const numCarrito = document.getElementById("num-carrito");
 
 discos.forEach((disc) => {
   const divProd = document.createElement("div");
@@ -176,7 +175,7 @@ discos.forEach((disc) => {
   <p>${disc.album}</p>
   <p>Año: ${disc.anio}</p>
   <h6>Precio: $${disc.precio}</h6>
-  <button type="button" class="btn btn-warning" onClick="agregarCarrito(${disc.id})">Agregar al carrito</button></div>`;
+  <button type="button" class="btn btn-warning" onClick="agregarCarrito(${disc.id})">Comprar <i class="fas fa-shopping-cart"></button></div>`;
   productos.append(divProd);
 });
 
@@ -185,7 +184,21 @@ const agregarCarrito = (id) => {
   const disc = discos.find((disco) => disco.id === id);
   carrito.push(disc);
   actCarrito();
+  
   console.log(carrito);
+
+  Toastify({
+    text: "Disco agregado al carrito",
+    duration: 1000,
+    style: {
+      background: "linear-gradient(to left, #ffb000, #764100)",
+    },
+    position: "center",
+    offset: {
+      x: 0,
+      y: 200,
+    },
+  }).showToast();
 };
 
 //funcion para borrar item agregado al carrito
@@ -194,6 +207,7 @@ const borrarItem = (id) => {
   const indice = carrito.indexOf(item);
   carrito.splice(indice, 1);
   actCarrito();
+  console.log(carrito)
 };
 
 const contCarrito = document.getElementById("contador-carrito");
@@ -208,22 +222,23 @@ const actCarrito = () => {
     mostProd.className = "row";
     mostProd.innerHTML = `
     <div class="col-6">  
-  <p>${disc.artista} - "${disc.album}"</p>
-  <hr>
-  </div> <div class="col-2">
-  <p> $${disc.precio}</p>
-  <hr>
-  </div><div class="col-3">
-  <p>1</p>
-  <hr>
-  </div>
-  <div class="col-1 btn-eliminar">
-  <button type="button" onClick="borrarItem(${disc.id})"><i class="fa-solid fa-trash-can"></i></button>
-  </div>`;
+    <p>${disc.artista} - "${disc.album}"</p>
+    <hr>
+    </div> <div class="col-3">
+    <p> $${disc.precio}</p>
+    <hr>
+    </div><div class="col-2">
+    <p>1</p>
+    <hr>
+    </div>
+    <div class="col-1 cont-eliminar">
+    <button type="button" class="btn-eliminar" onClick="borrarItem(${disc.id})"><i class="fa-solid fa-trash-can"></i></button>
+    </div>`;
     addCarrito.append(mostProd);
   });
 
   contCarrito.innerText = carrito.length;
+  numCarrito.innerText = carrito.length;
   total.innerText = carrito.reduce((ac, disc) => ac + disc.precio, 0);
 };
 
@@ -235,8 +250,26 @@ vaciar.addEventListener("click", (evt) => {
   console.log(carrito);
 });
 
-//--------------------------------------------------------------
+
+const modalCarrito = document.getElementById("modal-carrito");
+
+modalCarrito.addEventListener('click',(even)=>{
+  even.preventDefault();
+  actCarrito(); 
+});
+
+//--------------------------------------------------------------------------
 //calculo de pago en cuotas
+
+const calcOk = ()=>{
+  Swal.fire({
+    position: "center",
+    icon: "success",
+    title: "Bien hecho",
+    showConfirmButton: false,
+    timer: 1100,
+  });
+};
 
 let resultado;
 let cuota;
@@ -255,38 +288,48 @@ const montoInteres = (ev) => {
   if (cantCuotas === 3) {
     calculoInteres(monto, 0.05);
     cadaCuota(resultado, cantCuotas);
-    interes.innerText = resultado;
-    plan.innerText = cuota.toFixed(2);
+    interes.innerText = `$${resultado}`;
+    plan.innerText = `$${cuota.toFixed(2)}`;
     planCuota.innerText = ` en ${cantCuotas} cuotas`;
+    calcOk();
   } else if (cantCuotas === 6) {
     calculoInteres(monto, 0.1);
     cadaCuota(resultado, cantCuotas);
-    interes.innerText = resultado;
-    plan.innerText = cuota.toFixed(2);
+    interes.innerText = `$${resultado}`;
+    plan.innerText = `$${cuota.toFixed(2)}`;
     planCuota.innerText = ` en ${cantCuotas} cuotas`;
+    calcOk();
   } else if (cantCuotas === 12) {
     calculoInteres(monto, 0.15);
     cadaCuota(resultado, cantCuotas);
-    interes.innerText = resultado;
-    plan.innerText = cuota.toFixed(2);
+    interes.innerText = `$${resultado}`;
+    plan.innerText = `$${cuota.toFixed(2)}`;
     planCuota.innerText = ` en ${cantCuotas} cuotas`;
+    calcOk();
   } else if (cantCuotas === 18) {
     calculoInteres(monto, 0.2);
     cadaCuota(resultado, cantCuotas);
-    interes.innerText = resultado;
-    plan.innerText = cuota.toFixed(2);
+    interes.innerText = `$${resultado}`;
+    plan.innerText = `$${cuota.toFixed(2)}`;
     planCuota.innerText = ` en ${cantCuotas} cuotas`;
+    calcOk();
   } else {
-    interes.innerText = " NO CALCULADO";
-    plan.innerText = " NO CALCULADO";
+    interes.innerText = " No calculado";
+    plan.innerText = " No calculado";
     planCuota.innerText = ``;
     resultado = undefined;
     cuota = undefined;
+    Swal.fire({
+      icon: "error",
+      title: "Calculo no permitido!",
+      text: "Recuerde elegir entre 3, 6, 12 o 18 cuotas",
+    });
   }
 
   resultado
-    ? console.log(resultado)
-    : console.log("no se puede obtener un resultado");
+    ?
+    console.log(resultado) :
+    console.log("no se puede obtener un resultado");
 
   resultado ? console.log(cuota) : console.log("no se puede calcular cuotas");
